alerting/requirements.txt
pydantic>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
